Execute GQL operation after a demand has been made
Instead of executing the operation right after the subscription is instantiated, execute the operation after a demand is made. This avoids the crash where the subscription is technically not yet established but the subscriber might receive events. In such cases, a crash
```
Fatal error: Unexpected state: received completion but do not have subscription
```
can occur.